requirements.txt
h2==2.4.0
Twisted==16.3.0
pyOpenSSL==16.0.0
priority==1.1.1
service_identity
idna
